requirements.txt
tqdm
lz4
python-snappy
zstandard
numpy
pandas
xgboost
scikit-learn